+++ src/LibSharp/Caching/KeyValueCache.cs
@@ -25,11 +25,13 @@ public sealed class KeyValueCache<TKey, TValue> : IKeyValueCache<TKey, TValue>
     /// </summary>
     /// <param name="factory">Value factory.</param>
     /// <param name="timeToLive">Cache time-to-live.</param>
-    public KeyValueCache(Func<TKey, TValue> factory, TimeSpan timeToLive)
+    /// <param name="timeProvider">(Optional) Time provider used for expiration. Defaults to <see cref="TimeProvider.System"/>.</param>
+    public KeyValueCache(Func<TKey, TValue> factory, TimeSpan timeToLive, TimeProvider? timeProvider = null)
     {
         Argument.NotNull(factory, nameof(factory));
         Argument.GreaterThanOrEqualTo(timeToLive, TimeSpan.Zero, nameof(timeToLive));
 
+        m_timeProvider = timeProvider ?? TimeProvider.System;
         m_createFactory = factory;
         m_timeToLive = timeToLive;
     }
@@ -39,11 +41,13 @@ public sealed class KeyValueCache<TKey, TValue> : IKeyValueCache<TKey, TValue>
     /// </summary>
     /// <param name="factory">The value factory.</param>
     /// <param name="expirationFunction">Function to calculate expiration of a value.</param>
-    public KeyValueCache(Func<TKey, TValue> factory, Func<TKey, TValue, DateTime> expirationFunction)
+    /// <param name="timeProvider">(Optional) Time provider used for expiration. Defaults to <see cref="TimeProvider.System"/>.</param>
+    public KeyValueCache(Func<TKey, TValue> factory, Func<TKey, TValue, DateTime> expirationFunction, TimeProvider? timeProvider = null)
     {
         Argument.NotNull(factory, nameof(factory));
         Argument.NotNull(expirationFunction, nameof(expirationFunction));
 
+        m_timeProvider = timeProvider ?? TimeProvider.System;
         m_createFactory = factory;
         m_expirationFunction = expirationFunction;
     }
@@ -54,12 +58,14 @@ public sealed class KeyValueCache<TKey, TValue> : IKeyValueCache<TKey, TValue>
     /// <param name="createFactory">The creation factory.</param>
     /// <param name="updateFactory">The update factory.</param>
     /// <param name="timeToLive">Cache time-to-live.</param>
-    public KeyValueCache(Func<TKey, TValue> createFactory, Func<TKey, TValue, TValue> updateFactory, TimeSpan timeToLive)
+    /// <param name="timeProvider">(Optional) Time provider used for expiration. Defaults to <see cref="TimeProvider.System"/>.</param>
+    public KeyValueCache(Func<TKey, TValue> createFactory, Func<TKey, TValue, TValue> updateFactory, TimeSpan timeToLive, TimeProvider? timeProvider = null)
     {
         Argument.NotNull(createFactory, nameof(createFactory));
         Argument.NotNull(updateFactory, nameof(updateFactory));
         Argument.GreaterThanOrEqualTo(timeToLive, TimeSpan.Zero, nameof(timeToLive));
 
+        m_timeProvider = timeProvider ?? TimeProvider.System;
         m_createFactory = createFactory;
         m_updateFactory = updateFactory;
         m_timeToLive = timeToLive;
@@ -71,12 +77,14 @@ public sealed class KeyValueCache<TKey, TValue> : IKeyValueCache<TKey, TValue>
     /// <param name="createFactory">The creation factory.</param>
     /// <param name="updateFactory">The update factory.</param>
     /// <param name="expirationFunction">Function to calculate expiration of a value.</param>
-    public KeyValueCache(Func<TKey, TValue> createFactory, Func<TKey, TValue, TValue> updateFactory, Func<TKey, TValue, DateTime> expirationFunction)
+    /// <param name="timeProvider">(Optional) Time provider used for expiration. Defaults to <see cref="TimeProvider.System"/>.</param>
+    public KeyValueCache(Func<TKey, TValue> createFactory, Func<TKey, TValue, TValue> updateFactory, Func<TKey, TValue, DateTime> expirationFunction, TimeProvider? timeProvider = null)
     {
         Argument.NotNull(createFactory, nameof(createFactory));
         Argument.NotNull(updateFactory, nameof(updateFactory));
         Argument.NotNull(expirationFunction, nameof(expirationFunction));
 
+        m_timeProvider = timeProvider ?? TimeProvider.System;
         m_createFactory = createFactory;
         m_updateFactory = updateFactory;
         m_expirationFunction = expirationFunction;
@@ -101,16 +109,17 @@ public sealed class KeyValueCache<TKey, TValue> : IKeyValueCache<TKey, TValue>
         if (m_updateFactory is null)
         {
             return m_timeToLive.HasValue
-                ? new ValueCache<TValue>(() => m_createFactory(key), m_timeToLive.Value)
-                : new ValueCache<TValue>(() => m_createFactory(key), value => m_expirationFunction!(key, value));
+                ? new ValueCache<TValue>(() => m_createFactory(key), m_timeToLive.Value, m_timeProvider)
+                : new ValueCache<TValue>(() => m_createFactory(key), value => m_expirationFunction!(key, value), m_timeProvider);
         }
 
         return m_timeToLive.HasValue
-            ? new ValueCache<TValue>(() => m_createFactory(key), value => m_updateFactory(key, value), m_timeToLive.Value)
-            : new ValueCache<TValue>(() => m_createFactory(key), value => m_updateFactory(key, value), value => m_expirationFunction!(key, value));
+            ? new ValueCache<TValue>(() => m_createFactory(key), value => m_updateFactory(key, value), m_timeToLive.Value, m_timeProvider)
+            : new ValueCache<TValue>(() => m_createFactory(key), value => m_updateFactory(key, value), value => m_expirationFunction!(key, value), m_timeProvider);
     }
 
     private readonly ConcurrentDictionary<TKey, Lazy<ValueCache<TValue>>> m_cache = new ConcurrentDictionary<TKey, Lazy<ValueCache<TValue>>>();
+    private readonly TimeProvider m_timeProvider;
 
     // Exactly one of m_timeToLive / m_expirationFunction is set by each constructor; m_updateFactory
     // is null when the cache was created without an update factory. The forgiving access to